refactoring, copyright
