+++ casual/casual-api/src/main/java/se/laz/casual/config/ReverseInbound.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) 2022, The casual project. All rights reserved.
+ *
+ * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
+ */
 package se.laz.casual.config;
 
 import java.util.Objects;
