+++ casual/casual-network/src/main/java/se/laz/casual/network/inbound/reverse/StaggeredOptions.java
@@ -1,3 +1,9 @@
+/*
+ * Copyright (c) 2022, The casual project. All rights reserved.
+ *
+ * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
+ */
+
 package se.laz.casual.network.inbound.reverse;
 
 import java.time.Duration;
